podcast_summarizer/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
youtube-transcript-api
google-genai
